# RFC: use custom allocator for more block cache objects